srcs/requirements/selenium/requirements.txt
selenium
prometheus_client
bash
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
opentelemetry-exporter-otlp>=1.12.0
